Abandoned explicit indices and substitution.
Originally I had an idea for a VM that would use the indices and
substitutions in a clever way, but after thinking about the details for
a bit I think it is probably not as promising as I originally thought.

This commit replaces substitutions with more lisp like environments. It
is now much simpler and performs much faster.